// frontend/src/lib/api/generated/models/ListingTagsResponse.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { TagGroup } from '../models/TagGroup';
import { HttpFile } from '../http/http';

export class ListingTagsResponse {
    'data'?: Array<TagGroup>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<TagGroup>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ListingTagsResponse.attributeTypeMap;
    }

    public constructor() {
    }
}