// frontend/src/lib/api/generated/models/TagGroup.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { TagItem } from '../models/TagItem';
import { HttpFile } from '../http/http';

export class TagGroup {
    'groupId'?: number | null;
    'groupName'?: string | null;
    'tags'?: Array<TagItem>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "groupId",
            "baseName": "group_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "groupName",
            "baseName": "group_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "tags",
            "baseName": "tags",
            "type": "Array<TagItem>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return TagGroup.attributeTypeMap;
    }

    public constructor() {
    }
}